+++ Assets/Mansion/Scripts/Main/Controller/RoomItemDialogController.cs
@@ -33,7 +33,7 @@ public class RoomItemDialogController : DialogController {
 	}
 
 	public override void OnBuyButtonClicked () {
-		long keepMoneyCount = CountManager.Instance.KeepMoneyCount;
+		decimal keepMoneyCount = CountManager.Instance.KeepMoneyCount;
 		if(keepMoneyCount < PriceCalculator.CalcRoomItemPrice (mRoomData)){
 			//short money
 			SoundManager.Instance.PlaySE(AudioClipID.SE_SHORT_MONEY);
